Better handling of docker command (#15080)
Not all docker commands are replaced with functions now.

Earlier wer replaced all docker commands with a function to be able
to capture docker commands used and display it with -v for breeze.
This has proven to be harmful as
this is an unexpected behaviour for a docker command.

This change introduces docker_v command which outputs the command
when needed.

(cherry picked from commit 535e1a8e692ba28ad8ce9474a66b941af1df4875)