Better compatibility/diagnostics for arbitrary UID in docker image (#15162)
The PROD image of airflow is OpenShift compatible and it can be
run with either 'airflow' user (UID=50000) or with any other
user with (GID=0).

This change adds umask 0002 to make sure that whenever the image
is extended and new directories get created, the directories are
group-writeable for GID=0. This is added in the default
entrypoint.

The entrypoint will fail if it is not run as airflow user or if
other, arbitrary user is used with GID != 0.

Fixes: #15107
(cherry picked from commit ce91872eccceb8fb6277012a909ad6b529a071d2)